Reduce complexity and file size flagged by CodeScene
Extract _get_or_create_participant helper from _start_experiment_session to
reduce cyclomatic complexity. Split TestMetaCloudAPIServiceBSUIDRecipient
into a dedicated test file and move looks_like_bsuid unit tests to
test_datamodels.py where they belong.
