+++ apps/channels/tests/test_datamodels.py
@@ -1,6 +1,6 @@
 import pytest
 
-from apps.channels.datamodels import BaseMessage, MetaCloudAPIMessage, TwilioMessage
+from apps.channels.datamodels import BaseMessage, MetaCloudAPIMessage, TwilioMessage, looks_like_bsuid
 from apps.channels.models import ChannelPlatform
 from apps.channels.tests.message_examples import meta_cloud_api_messages
 
@@ -128,3 +128,29 @@ class TestTwilioMessageParse:
         }
         with pytest.raises(KeyError):
             TwilioMessage.parse(message)
+
+
+class TestLooksLikeBsuid:
+    """Unit tests for the ``looks_like_bsuid`` predicate."""
+
+    def test_bsuid_shapes_accepted(self):
+        """Meta's BSUID spec: ISO 3166 alpha-2 country code + period + up to 128 alphanumeric
+        characters. Parent BSUIDs (cross-portfolio) insert ``ENT.`` between country and identifier.
+        """
+        assert looks_like_bsuid("US.13491208655302741918") is True
+        assert looks_like_bsuid("US.ENT.11815799212886844830") is True
+        assert looks_like_bsuid("ZA.abc123XYZ") is True  # Alphanumeric tail, not just digits.
+        assert looks_like_bsuid("US." + "a" * 128) is True  # Spec maximum.
+
+    def test_non_bsuid_strings_are_rejected(self):
+        """Locale-formatted phones and other not-quite-BSUID strings must not pass."""
+        assert looks_like_bsuid("+1.212.555.2368") is False
+        assert looks_like_bsuid("us.13491208655302741918") is False  # Lowercase country.
+        assert looks_like_bsuid("USA.13491208655302741918") is False  # 3-letter country.
+        assert looks_like_bsuid("US.13491208655302741918.extra") is False  # Tail contains period.
+        assert looks_like_bsuid("US.has-dash") is False  # Non-alphanumeric in tail.
+        assert looks_like_bsuid("US." + "a" * 129) is False  # Tail exceeds 128 chars.
+        assert looks_like_bsuid("US.") is False  # Empty tail.
+        assert looks_like_bsuid("27456897512") is False  # Plain wa_id.
+        assert looks_like_bsuid("+27456897512") is False  # E.164.
+        assert looks_like_bsuid("") is False
