+++ apps/chat/channels.py
@@ -1451,6 +1451,44 @@ class CommCareConnectChannel(ChannelBase):
         return self.participant_data.get_encryption_key_bytes()
 
 
+def _get_or_create_participant(
+    team,
+    normalized_identifier: str,
+    platform,
+    participant_user: CustomUser | None,
+    participant_id_filter: Q,
+) -> Participant:
+    """Lookup or create a participant, handling complex (disjunctive) identity filters.
+
+    When participant_id_filter is a simple equality (the common case), delegates the
+    whole lookup-or-create to get_or_create. When it's a disjunction (e.g. BSUID OR
+    legacy phone), probes first so we can create with only the canonical identifier.
+    """
+    is_simple_filter = len(participant_id_filter.children) == 1
+    if not is_simple_filter:
+        existing = (
+            Participant.objects.filter(participant_id_filter, team=team, platform=platform)
+            .order_by("created_at")
+            .first()
+        )
+        if existing is not None:
+            if participant_user and existing.user is None:
+                existing.user = participant_user
+                existing.save()
+            return existing
+
+    participant, created = Participant.objects.get_or_create(
+        team=team,
+        identifier=normalized_identifier,
+        platform=platform,
+        defaults={"user": participant_user},
+    )
+    if not created and participant_user and participant.user is None:
+        participant.user = participant_user
+        participant.save()
+    return participant
+
+
 def _start_experiment_session(
     working_experiment: Experiment,
     experiment_channel: ExperimentChannel,
@@ -1480,31 +1518,13 @@ def _start_experiment_session(
         participant_id_filter = Q(identifier=normalized_identifier)
 
     with transaction.atomic():
-        # When participant_id_filter is a simple equality (the common case), delegate the
-        # whole lookup-or-create to get_or_create. When it's a disjunction (e.g. BSUID OR
-        # legacy phone), probe first so we can create with only the canonical identifier.
-        is_simple_filter = len(participant_id_filter.children) == 1
-        existing_participant = None
-        if not is_simple_filter:
-            matches = Participant.objects.filter(participant_id_filter, team=team, platform=experiment_channel.platform)
-            if matches.exists():
-                existing_participant = matches.order_by("created_at").first()
-
-        if existing_participant is None:
-            participant, created = Participant.objects.get_or_create(
-                team=team,
-                identifier=normalized_identifier,
-                platform=experiment_channel.platform,
-                defaults={"user": participant_user},
-            )
-            if not created and participant_user and participant.user is None:
-                participant.user = participant_user
-                participant.save()
-        else:
-            participant = existing_participant
-            if participant_user and participant.user is None:
-                participant.user = participant_user
-                participant.save()
+        participant = _get_or_create_participant(
+            team=team,
+            normalized_identifier=normalized_identifier,
+            platform=experiment_channel.platform,
+            participant_user=participant_user,
+            participant_id_filter=participant_id_filter,
+        )
 
         chat = Chat.objects.create(
             team=team,
@@ -1525,7 +1545,6 @@ def _start_experiment_session(
             },
         )
 
-        # Record the participant's timezone
         if timezone:
             participant.update_memory(data={"timezone": timezone}, experiment=working_experiment)
