+++ apps/service_providers/tests/test_messaging_providers.py
@@ -8,7 +8,7 @@ import pytest
 from django.utils import timezone
 from pydantic import ValidationError
 
-from apps.channels.datamodels import MetaCloudAPIMessage, TurnWhatsappMessage, looks_like_bsuid
+from apps.channels.datamodels import MetaCloudAPIMessage, TurnWhatsappMessage
 from apps.channels.models import ChannelPlatform
 from apps.channels.tests.message_examples import turnio_messages
 from apps.chat.channels import MESSAGE_TYPES
@@ -745,153 +745,6 @@ class TestMetaCloudAPIServiceWindow:
                 download_link="https://example.com/file",
                 last_activity_at=None,
             )
-
-
-class TestMetaCloudAPIServiceBSUIDRecipient:
-    """Outbound send paths must use Meta's `recipient` field when targeting a BSUID,
-    and the `to` field when targeting a phone number.
-
-    See https://developers.facebook.com/documentation/business-messaging/whatsapp/business-scoped-user-ids
-    """
-
-    def _make_service(self):
-        return MetaCloudAPIService(access_token="test_token", business_id="123456")
-
-    @patch("apps.service_providers.messaging_service.httpx.post")
-    def test_send_text_message_to_bsuid_uses_recipient_field(self, mock_post):
-        mock_post.return_value = httpx.Response(
-            200,
-            json={"messages": [{"id": "wamid.test"}]},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/messages"),
-        )
-        self._make_service().send_text_message(
-            message="Hello",
-            from_="phone123",
-            to="US.13491208655302741918",
-            platform=ChannelPlatform.WHATSAPP,
-            last_activity_at=timezone.now() - timedelta(hours=1),
-        )
-        sent = mock_post.call_args.kwargs["json"]
-        assert sent == {
-            "messaging_product": "whatsapp",
-            "recipient_type": "individual",
-            "recipient": "US.13491208655302741918",
-            "type": "text",
-            "text": {"body": "Hello"},
-        }
-        assert "to" not in sent
-
-    @patch("apps.service_providers.messaging_service.httpx.post")
-    def test_send_text_message_to_phone_keeps_to_field(self, mock_post):
-        mock_post.return_value = httpx.Response(
-            200,
-            json={"messages": [{"id": "wamid.test"}]},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/messages"),
-        )
-        self._make_service().send_text_message(
-            message="Hello",
-            from_="phone123",
-            to="+27826419977",
-            platform=ChannelPlatform.WHATSAPP,
-            last_activity_at=timezone.now() - timedelta(hours=1),
-        )
-        sent = mock_post.call_args.kwargs["json"]
-        assert sent["to"] == "+27826419977"
-        assert "recipient" not in sent
-
-    @patch("apps.service_providers.messaging_service.httpx.post")
-    def test_send_template_message_to_bsuid_uses_recipient_field(self, mock_post):
-        mock_post.return_value = httpx.Response(
-            200,
-            json={"messages": [{"id": "wamid.test"}]},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/messages"),
-        )
-        self._make_service().send_template_message(
-            message="Hi",
-            from_="phone123",
-            to="US.13491208655302741918",
-            platform=ChannelPlatform.WHATSAPP,
-        )
-        sent = mock_post.call_args.kwargs["json"]
-        assert sent["recipient"] == "US.13491208655302741918"
-        assert "to" not in sent
-
-    @patch("apps.service_providers.messaging_service.httpx.post")
-    def test_send_voice_message_to_bsuid_uses_recipient_field(self, mock_post):
-        upload_response = httpx.Response(
-            200,
-            json={"id": "media_id_abc"},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/media"),
-        )
-        send_response = httpx.Response(
-            200,
-            json={"messages": [{"id": "wamid.test"}]},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/messages"),
-        )
-        mock_post.side_effect = [upload_response, send_response]
-        synthetic_voice = MagicMock(spec=SynthesizedAudio)
-        synthetic_voice.get_audio_bytes.return_value = b"fake-ogg"
-        self._make_service().send_voice_message(
-            synthetic_voice=synthetic_voice,
-            from_="phone123",
-            to="US.13491208655302741918",
-            platform=ChannelPlatform.WHATSAPP,
-            last_activity_at=timezone.now() - timedelta(hours=1),
-        )
-        sent = mock_post.call_args_list[1].kwargs["json"]
-        assert sent["recipient"] == "US.13491208655302741918"
-        assert "to" not in sent
-
-    @patch("apps.service_providers.messaging_service.httpx.post")
-    def test_send_file_to_user_bsuid_uses_recipient_field(self, mock_post):
-        upload_response = httpx.Response(
-            200,
-            json={"id": "media_id_abc"},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/media"),
-        )
-        send_response = httpx.Response(
-            200,
-            json={"messages": [{"id": "wamid.test"}]},
-            request=httpx.Request("POST", "https://graph.facebook.com/v25.0/phone123/messages"),
-        )
-        mock_post.side_effect = [upload_response, send_response]
-        file = MagicMock()
-        file.content_type = "image/jpeg"
-        file.name = "x.jpg"
-        file.file.open.return_value.__enter__.return_value = BytesIO(b"data")
-        self._make_service().send_file_to_user(
-            from_="phone123",
-            to="US.13491208655302741918",
-            platform=ChannelPlatform.WHATSAPP,
-            file=file,
-            download_link="https://example.com/x.jpg",
-            last_activity_at=timezone.now() - timedelta(hours=1),
-        )
-        sent = mock_post.call_args_list[1].kwargs["json"]
-        assert sent["recipient"] == "US.13491208655302741918"
-        assert "to" not in sent
-
-    def test_bsuid_shapes_accepted(self):
-        """Meta's BSUID spec: ISO 3166 alpha-2 country code + period + up to 128 alphanumeric
-        characters. Parent BSUIDs (cross-portfolio) insert ``ENT.`` between country and identifier.
-        """
-        assert looks_like_bsuid("US.13491208655302741918") is True
-        assert looks_like_bsuid("US.ENT.11815799212886844830") is True
-        assert looks_like_bsuid("ZA.abc123XYZ") is True  # Alphanumeric tail, not just digits.
-        assert looks_like_bsuid("US." + "a" * 128) is True  # Spec maximum.
-
-    def test_non_bsuid_strings_are_rejected(self):
-        """Locale-formatted phones and other not-quite-BSUID strings must not pass."""
-        assert looks_like_bsuid("+1.212.555.2368") is False
-        assert looks_like_bsuid("us.13491208655302741918") is False  # Lowercase country.
-        assert looks_like_bsuid("USA.13491208655302741918") is False  # 3-letter country.
-        assert looks_like_bsuid("US.13491208655302741918.extra") is False  # Tail contains period.
-        assert looks_like_bsuid("US.has-dash") is False  # Non-alphanumeric in tail.
-        assert looks_like_bsuid("US." + "a" * 129) is False  # Tail exceeds 128 chars.
-        assert looks_like_bsuid("US.") is False  # Empty tail.
-        assert looks_like_bsuid("27456897512") is False  # Plain wa_id.
-        assert looks_like_bsuid("+27456897512") is False  # E.164.
-        assert looks_like_bsuid("") is False
 
 
 def _test_messaging_provider(team, provider_type: MessagingProviderType, data):
